requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.0.0
python-dotenv>=0.19.0
Pillow>=9.0.0
ttkthemes>=3.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
pandas>=1.5.0